Replaced jdk logging with log4j2 in standalone mode.
