--- browsermob-dist/src/main/config/bmp-logging.properties
@@ -1,22 +1,0 @@
-# specify the handlers to create in the root logger
-# (all loggers are children of the root logger)
-handlers=java.util.logging.ConsoleHandler
-    
-# set the default logging level for the root logger
-.level=INFO
-
-# to suppress unwanted logging statements, set the log level for the source logger to WARNING or SEVERE.
-# to enable more verbose logging, set the log level to FINE, FINER, or FINEST.
-net.lightbody.bmp=INFO
-
-# suppress some logging noise from Jetty 
-net.lightbody.bmp.proxy.jetty.util.ThreadedServer=WARNING
-
-# set the default logging level for new ConsoleHandler instances. 
-java.util.logging.ConsoleHandler.level=FINEST
-
-# set the default formatter for new ConsoleHandler instances
-java.util.logging.ConsoleHandler.formatter=java.util.logging.SimpleFormatter
-
-# logging format string. see http://docs.oracle.com/javase/7/docs/api/java/util/logging/SimpleFormatter.html for details.
-java.util.logging.SimpleFormatter.format=[%4$s %1$tF %1$tT.%1$tL %3$s] %5$s%n
