+++ browsermob-dist/src/main/resources/bmp-logging.yaml
@@ -1,0 +1,29 @@
+# This file controls the logging configuration for browsermob-proxy in "standalone" mode. To adjust the amount of log output, modify the
+# 'level' fields below. For more information on log4j configuration files, see http://logging.apache.org/log4j/2.x/manual/configuration.html.
+
+configuration:
+  name: standalone
+  appenders:
+    Console:
+      name: console
+      target: SYSTEM_OUT
+      PatternLayout:
+        pattern: "[%-5level %date{ISO8601} %logger] (%thread) %msg %xThrowable%n"
+  loggers:
+    logger:
+      -
+        name: net.lightbody.bmp.proxy.jetty.util.ThreadedServer
+        level: warn
+        additivity: false
+        appender-ref:
+          ref: console
+      -
+        name: net.lightbody.bmp
+        # to suppress unwanted BMP logging statements, set the level below for the source logger to WARN or ERROR.
+        # to enable more verbose logging, set the level to DEBUG or TRACE.
+        level: info
+    root:
+      # to suppress unwanted logging statements globally, set the level below to WARN or ERROR.
+      level: info
+      appender-ref:
+        ref: console
